backend/requirements.txt
# FastAPI 核心
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# 資料庫
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# 認證
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Socket.IO
python-socketio==5.11.0

# 驗證 & 工具
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0
